requirements.txt
numpy
opencv-contrib-python
opencv-python
openpyxl
pandas
Pillow
pyttsx3
